fix: correccion de estructuracion de prop TaskItem [Cruz Antonella, Efmimov Valentina, Padilla Santiago]
