+++ tp03PV2025Grupo21/src/assets/components/TaskList.jsx
@@ -1,23 +1,5 @@
-function TaskItem({ task }) {
-    return (
-      <li id={task?.id} className="todo_item">
-        <button className="items_check">
-          <svg>
-            <circle cx="11.998" cy="11.998" fillRule="nonzero" r="9.998" />
-          </svg>
-          <p>{task?.title}</p>
-        </button>
-        <div className="items_izquierda">
-          <button>
-            <span className="TextoOculto">Eliminar</span>
-            <svg>
-              <path d="" />
-            </svg>
-          </button>
-        </div>
-      </li>
-    );
-  } // define los elementos de la lista de tareas
+  import React from 'react';
+  import TaskItem from './TaskItem'; // Importa el componente TaskItem
   
   function TaskList({ tasks }) {
     return (
